[ADD] add previous PHP examples to this folder
There are several projects, practices, examples and other funny stuff
about PHP that we maybe will need one day, better to have all this in
one place.

Also, we are scavenging old memory sticks, dropbox, drive, and other
sources to place all the PHP we have in one place, so, this is a unique
commit, all PHP scavenging will be added here and no other
place(referring to a commit and folder).